Replaced cxcore code with the newer version from OpenCV 1.0.0.  This fixes some 64-bit compatibility bugs, and probably other bugs I don't know about.  The OpenCV version of cxcore uses "cv" as a prefix for type and function names, but this was changed to "cx" in the old PNL version.  Rather than making this change to the new cxcore code, I changed the rest of PNL to call cxcore functions with the cv prefix. This will make it easier to share cxcore revisions with OpenCV in the future.  The only parts of PNL that needed to be changed were the implementations of sparse matrices, decision trees, and random number generation.
With the new version of cxcore in place, I changed pnlSparseMatrix.hpp
to create CvMat objects with the correct element size for matrices of
pointers.  Previously it assumed that all objects stored in sparse
matrices were the same size as a 32-bit float, which is incorrect for
pointers on a 64-bit machine.  This fixed a number of "make check"
errors.

Switching to the new version of cxcore involved adding some source
files, which involved changing some Makefile.am files.  In addition to
checking in the new Makefile.am files, I'm also checking in new
Makefile.in files (generated by automake) and a new configure script
(generated from configure.in by autoconf).  These files are generated
by newer versions of automake (1.10 versus 1.8.2) and autoconf (2.61
versus 2.59) than the ones that were previously checked in.  Finally,
I removed the lines in configure.gcc and configure.icc that replace
all existing Makefile.in files with ones from an archive called
conf.tbz.  Since the Makefile.in files are included in the release and
in CVS, I don't see any reason to have conf.tbz, even for users who
don't have automake installed.